+++ batch/README.md
@@ -9,7 +9,7 @@
 $ export WEB3_HTTP_PROVIDER=http://localhost:8545
 $ export DATABASE_URL=postgresql://ethuser:ethpass@localhost:5432/ethcache
 $ export IBET_EXCHANGE_CONTRACT_ADDRESS=0x682e9123cb76b7842fb254c4c93dfce761e8faa4
-$ python async/indexer_DEX.py
+$ python batch/indexer_DEX.py
 ```
 
 ## processor_Notifications.py
