+++ batch/indexer_DEX.py
