+++ batch/lib/misc.py
