+++ batch/processor_Notifications_Coupon_Exchange.py
@@ -34,21 +34,21 @@ sys.path.append(path)
 from app import config
 from app.model import Notification, NotificationType
 from app.contracts import Contract
-from async.lib.token import TokenFactory
-from async.lib.company_list import CompanyListFactory
-from async.lib.token_list import TokenList
-from async.lib.misc import wait_all_futures
+from batch.lib.token import TokenFactory
+from batch.lib.company_list import CompanyListFactory
+from batch.lib.token_list import TokenList
+from batch.lib.misc import wait_all_futures
 import log
 
 JST = timezone(timedelta(hours=+9), "JST")
-LOG = log.get_logger(process_name="PROCESSOR-NOTIFICATIONS-SHARE-EXCHANGE")
+LOG = log.get_logger(process_name="PROCESSOR-NOTIFICATIONS-COUPON-EXCHANGE")
 
 # 設定の取得
 WEB3_HTTP_PROVIDER = config.WEB3_HTTP_PROVIDER
 URI = config.DATABASE_URL
 WORKER_COUNT = int(config.WORKER_COUNT)
 SLEEP_INTERVAL = int(config.SLEEP_INTERVAL)
-IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS = config.IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS
+IBET_CP_EXCHANGE_CONTRACT_ADDRESS = config.IBET_CP_EXCHANGE_CONTRACT_ADDRESS
 
 # 初期化
 web3 = Web3(Web3.HTTPProvider(WEB3_HTTP_PROVIDER))
@@ -63,7 +63,7 @@ company_list_factory = CompanyListFactory(config.COMPANY_LIST_URL)
 NOW_BLOCKNUMBER = web3.eth.blockNumber
 
 # コントラクトの生成
-share_exchange_contract = Contract.get_contract('IbetOTCExchange', IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS)
+cp_exchange_contract = Contract.get_contract('IbetCouponExchange', config.IBET_CP_EXCHANGE_CONTRACT_ADDRESS)
 list_contract = Contract.get_contract('TokenList', config.TOKEN_LIST_CONTRACT_ADDRESS)
 token_list = TokenList(list_contract)
 
@@ -76,9 +76,6 @@ class Watcher:
         self.filter_params = filter_params
         self.from_block = 0
 
-    def watch(self, entries):
-        pass
-
     def _gen_notification_id(self, entry, option_type=0):
         return "0x{:012x}{:06x}{:06x}{:02x}".format(
             entry["blockNumber"],
@@ -89,6 +86,9 @@ class Watcher:
 
     def _gen_block_timestamp(self, entry):
         return datetime.fromtimestamp(web3.eth.getBlock(entry["blockNumber"])["timestamp"], JST)
+
+    def watch(self, entries):
+        pass
 
     def loop(self):
         start_time = time.time()
@@ -126,14 +126,14 @@ class Watcher:
 
 
 '''
-株式取引（相対）関連（IbetOTCExchange）
+クーポン取引関連（IbetCouponExchange）
 '''
 
 
 # イベント：注文
-class WatchShareNewOrder(Watcher):
+class WatchCouponNewOrder(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "NewOrder", {})
+        super().__init__(cp_exchange_contract, "NewOrder", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -144,12 +144,7 @@ class WatchShareNewOrder(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
-
-            token = token_factory.get_share(token_address)
+            token = token_factory.get_coupon(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -157,25 +152,15 @@ class WatchShareNewOrder(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
             notification.notification_id = self._gen_notification_id(entry)
             notification.notification_type = NotificationType.NEW_ORDER.value
             notification.priority = 0
-            notification.address = entry["args"]["ownerAddress"]
-            notification.block_timestamp = self._gen_block_timestamp(entry)
-            notification.args = dict(entry["args"])
-            notification.metainfo = metadata
-            db_session.merge(notification)
-
-            notification = Notification()
-            notification.notification_id = self._gen_notification_id(entry)
-            notification.notification_type = NotificationType.NEW_ORDER_COUNTERPART.value
-            notification.priority = 0
-            notification.address = entry["args"]["counterpartAddress"]
+            notification.address = entry["args"]["accountAddress"]
             notification.block_timestamp = self._gen_block_timestamp(entry)
             notification.args = dict(entry["args"])
             notification.metainfo = metadata
@@ -183,9 +168,9 @@ class WatchShareNewOrder(Watcher):
 
 
 # イベント：注文取消
-class WatchShareCancelOrder(Watcher):
+class WatchCouponCancelOrder(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "CancelOrder", {})
+        super().__init__(cp_exchange_contract, "CancelOrder", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -196,38 +181,23 @@ class WatchShareCancelOrder(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
+            token = token_factory.get_coupon(token_address)
 
-            token = token_factory.get_share(token_address)
-            
             company = company_list.find(token.owner_address)
 
             metadata = {
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
             notification.notification_id = self._gen_notification_id(entry)
             notification.notification_type = NotificationType.CANCEL_ORDER.value
             notification.priority = 0
-            notification.address = entry["args"]["ownerAddress"]
-            notification.block_timestamp = self._gen_block_timestamp(entry)
-            notification.args = dict(entry["args"])
-            notification.metainfo = metadata
-            db_session.merge(notification)
-
-            notification = Notification()
-            notification.notification_id = self._gen_notification_id(entry)
-            notification.notification_type = NotificationType.CANCEL_ORDER_COUNTERPART.value
-            notification.priority = 0
-            notification.address = entry["args"]["counterpartAddress"]
+            notification.address = entry["args"]["accountAddress"]
             notification.block_timestamp = self._gen_block_timestamp(entry)
             notification.args = dict(entry["args"])
             notification.metainfo = metadata
@@ -235,9 +205,9 @@ class WatchShareCancelOrder(Watcher):
 
 
 # イベント：約定（買）
-class WatchShareBuyAgreement(Watcher):
+class WatchCouponBuyAgreement(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "Agree", {})
+        super().__init__(cp_exchange_contract, "Agree", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -248,21 +218,16 @@ class WatchShareBuyAgreement(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
+            token = token_factory.get_coupon(token_address)
 
-            token = token_factory.get_share(token_address)
-            
             company = company_list.find(token.owner_address)
 
             metadata = {
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
@@ -277,9 +242,9 @@ class WatchShareBuyAgreement(Watcher):
 
 
 # イベント：約定（売）
-class WatchShareSellAgreement(Watcher):
+class WatchCouponSellAgreement(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "Agree", {})
+        super().__init__(cp_exchange_contract, "Agree", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -290,12 +255,7 @@ class WatchShareSellAgreement(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
-
-            token = token_factory.get_share(token_address)
+            token = token_factory.get_coupon(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -303,8 +263,8 @@ class WatchShareSellAgreement(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
@@ -319,9 +279,9 @@ class WatchShareSellAgreement(Watcher):
 
 
 # イベント：決済OK（買）
-class WatchShareBuySettlementOK(Watcher):
+class WatchCouponBuySettlementOK(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "SettlementOK", {})
+        super().__init__(cp_exchange_contract, "SettlementOK", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -332,21 +292,16 @@ class WatchShareBuySettlementOK(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
+            token = token_factory.get_coupon(token_address)
 
-            token = token_factory.get_share(token_address)
-            
             company = company_list.find(token.owner_address)
 
             metadata = {
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
@@ -361,9 +316,9 @@ class WatchShareBuySettlementOK(Watcher):
 
 
 # イベント：決済OK（売）
-class WatchShareSellSettlementOK(Watcher):
+class WatchCouponSellSettlementOK(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "SettlementOK", {})
+        super().__init__(cp_exchange_contract, "SettlementOK", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -374,12 +329,7 @@ class WatchShareSellSettlementOK(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
-
-            token = token_factory.get_share(token_address)
+            token = token_factory.get_coupon(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -387,8 +337,8 @@ class WatchShareSellSettlementOK(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
@@ -403,9 +353,9 @@ class WatchShareSellSettlementOK(Watcher):
 
 
 # イベント：決済NG（買）
-class WatchShareBuySettlementNG(Watcher):
+class WatchCouponBuySettlementNG(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "SettlementNG", {})
+        super().__init__(cp_exchange_contract, "SettlementNG", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -416,12 +366,7 @@ class WatchShareBuySettlementNG(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
-
-            token = token_factory.get_share(token_address)
+            token = token_factory.get_coupon(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -429,8 +374,8 @@ class WatchShareBuySettlementNG(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
@@ -445,9 +390,9 @@ class WatchShareBuySettlementNG(Watcher):
 
 
 # イベント：決済NG（売）
-class WatchShareSellSettlementNG(Watcher):
+class WatchCouponSellSettlementNG(Watcher):
     def __init__(self):
-        super().__init__(share_exchange_contract, "SettlementNG", {})
+        super().__init__(cp_exchange_contract, "SettlementNG", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -458,12 +403,7 @@ class WatchShareSellSettlementNG(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
-            #       token_templateがShareではない場合処理をスキップする
-            if token_list.get_token(token_address)[1] != "IbetShare":
-                continue
-            
-            token = token_factory.get_share(token_address)
+            token = token_factory.get_coupon(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -471,8 +411,8 @@ class WatchShareSellSettlementNG(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetShare"
+                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetCoupon"
             }
 
             notification = Notification()
@@ -488,14 +428,14 @@ class WatchShareSellSettlementNG(Watcher):
 
 def main():
     watchers = [
-        WatchShareNewOrder(),
-        WatchShareCancelOrder(),
-        WatchShareBuyAgreement(),
-        WatchShareSellAgreement(),
-        WatchShareBuySettlementOK(),
-        WatchShareSellSettlementOK(),
-        WatchShareBuySettlementNG(),
-        WatchShareSellSettlementNG(),
+        WatchCouponNewOrder(),
+        WatchCouponCancelOrder(),
+        WatchCouponBuyAgreement(),
+        WatchCouponSellAgreement(),
+        WatchCouponBuySettlementOK(),
+        WatchCouponSellSettlementOK(),
+        WatchCouponBuySettlementNG(),
+        WatchCouponSellSettlementNG(),
     ]
 
     e = ThreadPoolExecutor(max_workers=WORKER_COUNT)
