+++ batch/processor_Notifications_Membership_Exchange.py
@@ -34,21 +34,21 @@ sys.path.append(path)
 from app import config
 from app.model import Notification, NotificationType
 from app.contracts import Contract
-from async.lib.token import TokenFactory
-from async.lib.company_list import CompanyListFactory
-from async.lib.token_list import TokenList
-from async.lib.misc import wait_all_futures
+from batch.lib.token import TokenFactory
+from batch.lib.company_list import CompanyListFactory
+from batch.lib.token_list import TokenList
+from batch.lib.misc import wait_all_futures
 import log
 
 JST = timezone(timedelta(hours=+9), "JST")
-LOG = log.get_logger(process_name="PROCESSOR-NOTIFICATIONS-COUPON-EXCHANGE")
+LOG = log.get_logger(process_name="PROCESSOR-NOTIFICATIONS-MEMBERSHIP-EXCHANGE")
 
 # 設定の取得
 WEB3_HTTP_PROVIDER = config.WEB3_HTTP_PROVIDER
 URI = config.DATABASE_URL
 WORKER_COUNT = int(config.WORKER_COUNT)
 SLEEP_INTERVAL = int(config.SLEEP_INTERVAL)
-IBET_CP_EXCHANGE_CONTRACT_ADDRESS = config.IBET_CP_EXCHANGE_CONTRACT_ADDRESS
+IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS = config.IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS
 
 # 初期化
 web3 = Web3(Web3.HTTPProvider(WEB3_HTTP_PROVIDER))
@@ -63,7 +63,10 @@ company_list_factory = CompanyListFactory(config.COMPANY_LIST_URL)
 NOW_BLOCKNUMBER = web3.eth.blockNumber
 
 # コントラクトの生成
-cp_exchange_contract = Contract.get_contract('IbetCouponExchange', config.IBET_CP_EXCHANGE_CONTRACT_ADDRESS)
+membership_exchange_contract = Contract.get_contract(
+    'IbetMembershipExchange',
+    config.IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS
+)
 list_contract = Contract.get_contract('TokenList', config.TOKEN_LIST_CONTRACT_ADDRESS)
 token_list = TokenList(list_contract)
 
@@ -126,14 +129,14 @@ class Watcher:
 
 
 '''
-クーポン取引関連（IbetCouponExchange）
+会員権取引関連（IbetMembershipExchange）
 '''
 
 
 # イベント：注文
-class WatchCouponNewOrder(Watcher):
+class WatchMembershipNewOrder(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "NewOrder", {})
+        super().__init__(membership_exchange_contract, "NewOrder", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -144,7 +147,7 @@ class WatchCouponNewOrder(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -152,8 +155,8 @@ class WatchCouponNewOrder(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -168,9 +171,9 @@ class WatchCouponNewOrder(Watcher):
 
 
 # イベント：注文取消
-class WatchCouponCancelOrder(Watcher):
+class WatchMembershipCancelOrder(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "CancelOrder", {})
+        super().__init__(membership_exchange_contract, "CancelOrder", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -181,7 +184,7 @@ class WatchCouponCancelOrder(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -189,8 +192,8 @@ class WatchCouponCancelOrder(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -205,9 +208,9 @@ class WatchCouponCancelOrder(Watcher):
 
 
 # イベント：約定（買）
-class WatchCouponBuyAgreement(Watcher):
+class WatchMembershipBuyAgreement(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "Agree", {})
+        super().__init__(membership_exchange_contract, "Agree", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -218,7 +221,7 @@ class WatchCouponBuyAgreement(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -226,8 +229,8 @@ class WatchCouponBuyAgreement(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -242,9 +245,9 @@ class WatchCouponBuyAgreement(Watcher):
 
 
 # イベント：約定（売）
-class WatchCouponSellAgreement(Watcher):
+class WatchMembershipSellAgreement(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "Agree", {})
+        super().__init__(membership_exchange_contract, "Agree", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -255,7 +258,7 @@ class WatchCouponSellAgreement(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -263,8 +266,8 @@ class WatchCouponSellAgreement(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -279,9 +282,9 @@ class WatchCouponSellAgreement(Watcher):
 
 
 # イベント：決済OK（買）
-class WatchCouponBuySettlementOK(Watcher):
+class WatchMembershipBuySettlementOK(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "SettlementOK", {})
+        super().__init__(membership_exchange_contract, "SettlementOK", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -292,7 +295,7 @@ class WatchCouponBuySettlementOK(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -300,8 +303,8 @@ class WatchCouponBuySettlementOK(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -316,9 +319,9 @@ class WatchCouponBuySettlementOK(Watcher):
 
 
 # イベント：決済OK（売）
-class WatchCouponSellSettlementOK(Watcher):
+class WatchMembershipSellSettlementOK(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "SettlementOK", {})
+        super().__init__(membership_exchange_contract, "SettlementOK", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -329,7 +332,7 @@ class WatchCouponSellSettlementOK(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -337,8 +340,8 @@ class WatchCouponSellSettlementOK(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -353,9 +356,9 @@ class WatchCouponSellSettlementOK(Watcher):
 
 
 # イベント：決済NG（買）
-class WatchCouponBuySettlementNG(Watcher):
+class WatchMembershipBuySettlementNG(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "SettlementNG", {})
+        super().__init__(membership_exchange_contract, "SettlementNG", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -366,7 +369,7 @@ class WatchCouponBuySettlementNG(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -374,8 +377,8 @@ class WatchCouponBuySettlementNG(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -390,9 +393,9 @@ class WatchCouponBuySettlementNG(Watcher):
 
 
 # イベント：決済NG（売）
-class WatchCouponSellSettlementNG(Watcher):
+class WatchMembershipSellSettlementNG(Watcher):
     def __init__(self):
-        super().__init__(cp_exchange_contract, "SettlementNG", {})
+        super().__init__(membership_exchange_contract, "SettlementNG", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -403,7 +406,7 @@ class WatchCouponSellSettlementNG(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_coupon(token_address)
+            token = token_factory.get_membership(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -411,8 +414,8 @@ class WatchCouponSellSettlementNG(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_CP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetCoupon"
+                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetMembership"
             }
 
             notification = Notification()
@@ -428,14 +431,14 @@ class WatchCouponSellSettlementNG(Watcher):
 
 def main():
     watchers = [
-        WatchCouponNewOrder(),
-        WatchCouponCancelOrder(),
-        WatchCouponBuyAgreement(),
-        WatchCouponSellAgreement(),
-        WatchCouponBuySettlementOK(),
-        WatchCouponSellSettlementOK(),
-        WatchCouponBuySettlementNG(),
-        WatchCouponSellSettlementNG(),
+        WatchMembershipNewOrder(),
+        WatchMembershipCancelOrder(),
+        WatchMembershipBuyAgreement(),
+        WatchMembershipSellAgreement(),
+        WatchMembershipBuySettlementOK(),
+        WatchMembershipSellSettlementOK(),
+        WatchMembershipBuySettlementNG(),
+        WatchMembershipSellSettlementNG(),
     ]
 
     e = ThreadPoolExecutor(max_workers=WORKER_COUNT)
