+++ batch/processor_Notifications_Share_Exchange.py
@@ -34,21 +34,21 @@ sys.path.append(path)
 from app import config
 from app.model import Notification, NotificationType
 from app.contracts import Contract
-from async.lib.token import TokenFactory
-from async.lib.company_list import CompanyListFactory
-from async.lib.token_list import TokenList
-from async.lib.misc import wait_all_futures
+from batch.lib.token import TokenFactory
+from batch.lib.company_list import CompanyListFactory
+from batch.lib.token_list import TokenList
+from batch.lib.misc import wait_all_futures
 import log
 
 JST = timezone(timedelta(hours=+9), "JST")
-LOG = log.get_logger(process_name="PROCESSOR-NOTIFICATIONS-MEMBERSHIP-EXCHANGE")
+LOG = log.get_logger(process_name="PROCESSOR-NOTIFICATIONS-SHARE-EXCHANGE")
 
 # 設定の取得
 WEB3_HTTP_PROVIDER = config.WEB3_HTTP_PROVIDER
 URI = config.DATABASE_URL
 WORKER_COUNT = int(config.WORKER_COUNT)
 SLEEP_INTERVAL = int(config.SLEEP_INTERVAL)
-IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS = config.IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS
+IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS = config.IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS
 
 # 初期化
 web3 = Web3(Web3.HTTPProvider(WEB3_HTTP_PROVIDER))
@@ -63,10 +63,7 @@ company_list_factory = CompanyListFactory(config.COMPANY_LIST_URL)
 NOW_BLOCKNUMBER = web3.eth.blockNumber
 
 # コントラクトの生成
-membership_exchange_contract = Contract.get_contract(
-    'IbetMembershipExchange',
-    config.IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS
-)
+share_exchange_contract = Contract.get_contract('IbetOTCExchange', IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS)
 list_contract = Contract.get_contract('TokenList', config.TOKEN_LIST_CONTRACT_ADDRESS)
 token_list = TokenList(list_contract)
 
@@ -79,6 +76,9 @@ class Watcher:
         self.filter_params = filter_params
         self.from_block = 0
 
+    def watch(self, entries):
+        pass
+
     def _gen_notification_id(self, entry, option_type=0):
         return "0x{:012x}{:06x}{:06x}{:02x}".format(
             entry["blockNumber"],
@@ -89,9 +89,6 @@ class Watcher:
 
     def _gen_block_timestamp(self, entry):
         return datetime.fromtimestamp(web3.eth.getBlock(entry["blockNumber"])["timestamp"], JST)
-
-    def watch(self, entries):
-        pass
 
     def loop(self):
         start_time = time.time()
@@ -129,14 +126,14 @@ class Watcher:
 
 
 '''
-会員権取引関連（IbetMembershipExchange）
+株式取引（相対）関連（IbetOTCExchange）
 '''
 
 
 # イベント：注文
-class WatchMembershipNewOrder(Watcher):
+class WatchShareNewOrder(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "NewOrder", {})
+        super().__init__(share_exchange_contract, "NewOrder", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -147,7 +144,12 @@ class WatchMembershipNewOrder(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
+
+            token = token_factory.get_share(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -155,15 +157,25 @@ class WatchMembershipNewOrder(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
             notification.notification_id = self._gen_notification_id(entry)
             notification.notification_type = NotificationType.NEW_ORDER.value
             notification.priority = 0
-            notification.address = entry["args"]["accountAddress"]
+            notification.address = entry["args"]["ownerAddress"]
+            notification.block_timestamp = self._gen_block_timestamp(entry)
+            notification.args = dict(entry["args"])
+            notification.metainfo = metadata
+            db_session.merge(notification)
+
+            notification = Notification()
+            notification.notification_id = self._gen_notification_id(entry)
+            notification.notification_type = NotificationType.NEW_ORDER_COUNTERPART.value
+            notification.priority = 0
+            notification.address = entry["args"]["counterpartAddress"]
             notification.block_timestamp = self._gen_block_timestamp(entry)
             notification.args = dict(entry["args"])
             notification.metainfo = metadata
@@ -171,9 +183,9 @@ class WatchMembershipNewOrder(Watcher):
 
 
 # イベント：注文取消
-class WatchMembershipCancelOrder(Watcher):
+class WatchShareCancelOrder(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "CancelOrder", {})
+        super().__init__(share_exchange_contract, "CancelOrder", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -184,23 +196,38 @@ class WatchMembershipCancelOrder(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
 
+            token = token_factory.get_share(token_address)
+            
             company = company_list.find(token.owner_address)
 
             metadata = {
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
             notification.notification_id = self._gen_notification_id(entry)
             notification.notification_type = NotificationType.CANCEL_ORDER.value
             notification.priority = 0
-            notification.address = entry["args"]["accountAddress"]
+            notification.address = entry["args"]["ownerAddress"]
+            notification.block_timestamp = self._gen_block_timestamp(entry)
+            notification.args = dict(entry["args"])
+            notification.metainfo = metadata
+            db_session.merge(notification)
+
+            notification = Notification()
+            notification.notification_id = self._gen_notification_id(entry)
+            notification.notification_type = NotificationType.CANCEL_ORDER_COUNTERPART.value
+            notification.priority = 0
+            notification.address = entry["args"]["counterpartAddress"]
             notification.block_timestamp = self._gen_block_timestamp(entry)
             notification.args = dict(entry["args"])
             notification.metainfo = metadata
@@ -208,9 +235,9 @@ class WatchMembershipCancelOrder(Watcher):
 
 
 # イベント：約定（買）
-class WatchMembershipBuyAgreement(Watcher):
+class WatchShareBuyAgreement(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "Agree", {})
+        super().__init__(share_exchange_contract, "Agree", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -221,16 +248,21 @@ class WatchMembershipBuyAgreement(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
 
+            token = token_factory.get_share(token_address)
+            
             company = company_list.find(token.owner_address)
 
             metadata = {
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
@@ -245,9 +277,9 @@ class WatchMembershipBuyAgreement(Watcher):
 
 
 # イベント：約定（売）
-class WatchMembershipSellAgreement(Watcher):
+class WatchShareSellAgreement(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "Agree", {})
+        super().__init__(share_exchange_contract, "Agree", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -258,7 +290,12 @@ class WatchMembershipSellAgreement(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
+
+            token = token_factory.get_share(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -266,8 +303,8 @@ class WatchMembershipSellAgreement(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
@@ -282,9 +319,9 @@ class WatchMembershipSellAgreement(Watcher):
 
 
 # イベント：決済OK（買）
-class WatchMembershipBuySettlementOK(Watcher):
+class WatchShareBuySettlementOK(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "SettlementOK", {})
+        super().__init__(share_exchange_contract, "SettlementOK", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -295,16 +332,21 @@ class WatchMembershipBuySettlementOK(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
 
+            token = token_factory.get_share(token_address)
+            
             company = company_list.find(token.owner_address)
 
             metadata = {
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
@@ -319,9 +361,9 @@ class WatchMembershipBuySettlementOK(Watcher):
 
 
 # イベント：決済OK（売）
-class WatchMembershipSellSettlementOK(Watcher):
+class WatchShareSellSettlementOK(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "SettlementOK", {})
+        super().__init__(share_exchange_contract, "SettlementOK", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -332,7 +374,12 @@ class WatchMembershipSellSettlementOK(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
+
+            token = token_factory.get_share(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -340,8 +387,8 @@ class WatchMembershipSellSettlementOK(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
@@ -356,9 +403,9 @@ class WatchMembershipSellSettlementOK(Watcher):
 
 
 # イベント：決済NG（買）
-class WatchMembershipBuySettlementNG(Watcher):
+class WatchShareBuySettlementNG(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "SettlementNG", {})
+        super().__init__(share_exchange_contract, "SettlementNG", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -369,7 +416,12 @@ class WatchMembershipBuySettlementNG(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
+
+            token = token_factory.get_share(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -377,8 +429,8 @@ class WatchMembershipBuySettlementNG(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
@@ -393,9 +445,9 @@ class WatchMembershipBuySettlementNG(Watcher):
 
 
 # イベント：決済NG（売）
-class WatchMembershipSellSettlementNG(Watcher):
+class WatchShareSellSettlementNG(Watcher):
     def __init__(self):
-        super().__init__(membership_exchange_contract, "SettlementNG", {})
+        super().__init__(share_exchange_contract, "SettlementNG", {})
 
     def watch(self, entries):
         company_list = company_list_factory.get()
@@ -406,7 +458,12 @@ class WatchMembershipSellSettlementNG(Watcher):
             if not token_list.is_registered(token_address):
                 continue
 
-            token = token_factory.get_membership(token_address)
+            # NOTE: OTCExchangeはShare以外のトークンでも利用される可能性があるため、
+            #       token_templateがShareではない場合処理をスキップする
+            if token_list.get_token(token_address)[1] != "IbetShare":
+                continue
+            
+            token = token_factory.get_share(token_address)
 
             company = company_list.find(token.owner_address)
 
@@ -414,8 +471,8 @@ class WatchMembershipSellSettlementNG(Watcher):
                 "company_name": company.corporate_name,
                 "token_address": token_address,
                 "token_name": token.name,
-                "exchange_address": IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS,
-                "token_type": "IbetMembership"
+                "exchange_address": IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS,
+                "token_type": "IbetShare"
             }
 
             notification = Notification()
@@ -431,14 +488,14 @@ class WatchMembershipSellSettlementNG(Watcher):
 
 def main():
     watchers = [
-        WatchMembershipNewOrder(),
-        WatchMembershipCancelOrder(),
-        WatchMembershipBuyAgreement(),
-        WatchMembershipSellAgreement(),
-        WatchMembershipBuySettlementOK(),
-        WatchMembershipSellSettlementOK(),
-        WatchMembershipBuySettlementNG(),
-        WatchMembershipSellSettlementNG(),
+        WatchShareNewOrder(),
+        WatchShareCancelOrder(),
+        WatchShareBuyAgreement(),
+        WatchShareSellAgreement(),
+        WatchShareBuySettlementOK(),
+        WatchShareSellSettlementOK(),
+        WatchShareBuySettlementNG(),
+        WatchShareSellSettlementNG(),
     ]
 
     e = ThreadPoolExecutor(max_workers=WORKER_COUNT)
