+++ bin/healthcheck_indexer.sh
@@ -17,30 +17,30 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-PROC_LIST="${PROC_LIST} async/indexer_Transfer.py"
+PROC_LIST="${PROC_LIST} batch/indexer_Transfer.py"
 
 if [ "${SHARE_TOKEN_ENABLED}" = 1 ]; then
-  PROC_LIST="${PROC_LIST} async/indexer_Position_Share.py"
+  PROC_LIST="${PROC_LIST} batch/indexer_Position_Share.py"
 fi
 
 if [ "${BOND_TOKEN_ENABLED}" = 1 ]; then
-  PROC_LIST="${PROC_LIST} async/indexer_Position_Bond.py"
+  PROC_LIST="${PROC_LIST} batch/indexer_Position_Bond.py"
 fi
 
 if [ "${MEMBERSHIP_TOKEN_ENABLED}" = 1 ]; then
-  PROC_LIST="${PROC_LIST} async/indexer_Position_Membership.py"
+  PROC_LIST="${PROC_LIST} batch/indexer_Position_Membership.py"
 fi
 
 if [ "${COUPON_TOKEN_ENABLED}" = 1 ]; then
-  PROC_LIST="${PROC_LIST} async/indexer_Consume_Coupon.py"
-  PROC_LIST="${PROC_LIST} async/indexer_Position_Coupon.py"
+  PROC_LIST="${PROC_LIST} batch/indexer_Consume_Coupon.py"
+  PROC_LIST="${PROC_LIST} batch/indexer_Position_Coupon.py"
 fi
 
 if [ -n "$IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS" -o \
   -n "$IBET_SB_EXCHANGE_CONTRACT_ADDRESS" -o \
   -n "$IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS" -o \
   -n "$IBET_CP_EXCHANGE_CONTRACT_ADDRESS" ]; then
-  PROC_LIST="${PROC_LIST} async/indexer_DEX.py"
+  PROC_LIST="${PROC_LIST} batch/indexer_DEX.py"
 fi
 
 for i in ${PROC_LIST}; do
