+++ bin/run_processor_notification.sh
@@ -23,29 +23,29 @@ source ~/.bash_profile
 cd /app/ibet-Wallet-API
 
 if [ "$BOND_TOKEN_ENABLED" = 1 ]; then
-  python async/processor_Notifications_Bond_Token.py &
+  python batch/processor_Notifications_Bond_Token.py &
   if [ ! -z "${IBET_SB_EXCHANGE_CONTRACT_ADDRESS}" ]; then
-    python async/processor_Notifications_Bond_Exchange.py &
+    python batch/processor_Notifications_Bond_Exchange.py &
   fi
 fi
 
 if [ "$SHARE_TOKEN_ENABLED" = 1 ]; then
-  python async/processor_Notifications_Share_Token.py &
+  python batch/processor_Notifications_Share_Token.py &
   if [ ! -z "${IBET_SHARE_EXCHANGE_CONTRACT_ADDRESS}" ]; then
-    python async/processor_Notifications_Share_Exchange.py &
+    python batch/processor_Notifications_Share_Exchange.py &
   fi
 fi
 
 if [ "$MEMBERSHIP_TOKEN_ENABLED" = 1 ]; then
   if [ ! -z "${IBET_MEMBERSHIP_EXCHANGE_CONTRACT_ADDRESS}" ]; then
-    python async/processor_Notifications_Membership_Exchange.py &
+    python batch/processor_Notifications_Membership_Exchange.py &
   fi
 fi
 
 if [ "$COUPON_TOKEN_ENABLED" = 1 ]; then
-  python async/processor_Notifications_Coupon_Token.py &
+  python batch/processor_Notifications_Coupon_Token.py &
   if [ ! -z "${IBET_CP_EXCHANGE_CONTRACT_ADDRESS}" ]; then
-    python async/processor_Notifications_Coupon_Exchange.py &
+    python batch/processor_Notifications_Coupon_Exchange.py &
   fi
 fi
