+++ requirements.txt
@@ -1,37 +1,14 @@
-asn1crypto==0.24.0
-attrdict==2.0.0
 Cerberus==0.9
-cffi==1.11.4
-coincurve==7.1.0
-eth-account==0.2.2
-eth-keyfile==0.5.0
-eth-rlp==0.1.2
-eth-utils==1.9.5
-ethereum==2.3.0
+coincurve==14.0.0
+eth-keyfile==0.5.1
+eth-keys<0.3.0
 falcon==1.4.1
 future==0.16.0
 gunicorn==20.0.4
 gevent==1.4.0
 hexbytes==0.1.0
-lru-dict==1.1.6
-parsimonious==0.8.0
-pbkdf2==1.3
+PyMySQL==0.10.1
 psycopg2-binary==2.8.6
-py-ecc==1.4.2
-py-solc==2.1.0
-pycparser==2.18
-pycryptodome==3.6.6
-pyethash==0.1.27
-pysha3==1.0.2
-python-mimeparse==1.6.0
-PyYAML==5.4
-repoze.lru==0.7
-rlp==0.6.0
-scrypt==0.8.6
-semantic-version==2.6.0
-six==1.12.0
 SQLAlchemy>=1.3.0
 sqlalchemy-migrate==0.13.0
-toolz==0.9.0
 web3==4.10.0
-PyMySQL==0.10.1
