+++ runtime.txt
@@ -1,1 +1,1 @@
-python-3.6.8
+python-3.8.7
